Transfere tradutor de preprints para módulo translator/*.py
