+++ scielo_usage_counter/url_translator.py
@@ -49,14 +49,6 @@ MEDIA_FORMAT_PDF = 'pdf'
 MEDIA_FORMAT_XML = 'xml'
 MEDIA_FORMAT_UNDEFINED = 'und'
 
-# Patterns related to Preprints site
-REGEX_PREPRINTS_SITE_VIEW_ABSTRACT = re.compile(r'preprint/view/(?P<id>\d+)$', re.IGNORECASE)
-REGEX_PREPRINTS_SITE_DOCUMENT_ABSTRACT = re.compile(r'documents/article/view/(?P<id>\d+)$', re.IGNORECASE)
-REGEX_PREPRINTS_SITE_VERSION_ABSTRACT = re.compile(r'preprint/view/(\d+)/version/(?P<id>\d+)$', re.IGNORECASE)
-REGEX_PREPRINTS_SITE_VIEW_PDF = re.compile(r'preprint/view/(?P<id>\d+)/(\d+)', re.IGNORECASE)
-REGEX_PREPRINTS_SITE_DOWNLOAD_PDF = re.compile(r'preprint/download/(?P<id>\d+)/(\d+)', re.IGNORECASE)
-REGEX_PREPRINTS_SITE_DOCUMENT_DOWNLOAD_PDF = re.compile(r'documents/article/download/(?P<id>\d+)/(\d+)', re.IGNORECASE)
-REGEX_PREPRINTS_SITE_VERSION_DOWNLOAD_PDF = re.compile(r'preprint/download/(?P<id>\d+)/version/(\d+)/(\d+)', re.IGNORECASE)
 
 # Patterns related to OPAC Alpha site
 REGEX_OPAC_ALPHA_JOURNAL_ARTICLE_HTML_DETAILS = re.compile(r'.*/article/(\w*)/?([\d|\w|\.|\-]*)/?([\d|\w|\-]*)/?(\w*|)', re.IGNORECASE)
@@ -188,187 +180,6 @@ class URLTranslationManager:
                     std_fields[k] = ''            
 
         return std_fields
-
-
-class URLTranslatorPreprintsSite:
-    def __init__(self, journals_metadata, articles_metadata):
-        self.journals_metadata = journals_metadata
-        self.articles_metadata = articles_metadata
-        self.name = NAME_PREPRINTS_SITE
-
-    def pipeline_translate(self, url):
-        parsed_url = urlparse(url)
-        query_params = parse_qs(parsed_url.query)
-
-        preprint_id = self.extract_preprint_id(parsed_url, query_params)
-        scielo_issn = self.extract_issn(preprint_id)        
-        media_format = self.extract_media_format(parsed_url, query_params)
-        media_language = self.extract_media_language(preprint_id)
-
-        return {
-            'scielo_issn': scielo_issn,
-            'pid_v2': preprint_id,
-            'pid_v3': None,
-            'media_format': media_format,
-            'media_language': media_language,
-        }
-
-    @classmethod
-    def extract_media_format(cls, url):
-        unquoted_url = urlparse.unquote(url)
-
-        html_patterns = [
-            REGEX_PREPRINTS_SITE_VIEW_ABSTRACT,
-            REGEX_PREPRINTS_SITE_DOCUMENT_ABSTRACT,
-            REGEX_PREPRINTS_SITE_VERSION_ABSTRACT
-        ]
-
-        pdf_patterns = [
-            REGEX_PREPRINTS_SITE_VIEW_PDF,
-            REGEX_PREPRINTS_SITE_DOWNLOAD_PDF,
-            REGEX_PREPRINTS_SITE_DOCUMENT_DOWNLOAD_PDF,
-            REGEX_PREPRINTS_SITE_VERSION_DOWNLOAD_PDF
-        ]
-
-        if any(re.search(pattern, unquoted_url) for pattern in html_patterns):
-            return MEDIA_FORMAT_HTML
-
-        if any(re.search(pattern, unquoted_url) for pattern in pdf_patterns):
-            return MEDIA_FORMAT_PDF
-
-        return MEDIA_FORMAT_UNDEFINED
-
-    def extract_media_language(self, pid_v2):
-        return self.articles_metadata['pid_v2_to_default_lang'].get(pid_v2)
-
-    @classmethod
-    def extract_issn():
-        return DEFAULT_SCIELO_ISSN
-
-    @classmethod
-    def extract_preprint_id(url):
-        unquoted_url = urlparse.unquote(url)
-
-        for pattern in [
-            REGEX_PREPRINTS_SITE_VIEW_ABSTRACT,
-            REGEX_PREPRINTS_SITE_DOCUMENT_ABSTRACT,
-            REGEX_PREPRINTS_SITE_VERSION_ABSTRACT,
-            REGEX_PREPRINTS_SITE_VIEW_PDF,
-            REGEX_PREPRINTS_SITE_DOWNLOAD_PDF,
-            REGEX_PREPRINTS_SITE_DOCUMENT_DOWNLOAD_PDF,
-            REGEX_PREPRINTS_SITE_VERSION_DOWNLOAD_PDF
-        ]:
-            match = re.search(pattern, unquoted_url)
-            if match:
-                return match.group('id')
-
-
-class URLTranslatorOPACAlphaSite:
-    def __init__(self, journals_metadata, articles_metadata):
-        self.journals_metadata = journals_metadata
-        self.articles_metadata = articles_metadata
-        self.name = NAME_OPAC_ALPHA_SITE
-
-    def pipeline_translate(self, url):
-        self.url_params = self.extract_url_params(url)
-
-        pid_v3 = self.extract_pid_v3(url)
-        media_format = self.extract_media_format(url)
-        media_language = self.extract_media_language(pid_v3)
-        scielo_issn = self.extract_issn(url)
-
-        return {
-            'scielo_issn': scielo_issn,
-            'pid_v2': None,
-            'pid_v3': pid_v3,
-            'media_format': media_format,
-            'media_language': media_language,
-        }
-
-    def extract_url_params(self, url):
-        url_params = {
-            'pid_v3': '', 
-            'journal_acronym': '',
-            'media_format': '',
-            'media_language': '',
-            'resource_ssm_path': ''
-        }
-
-        url_evaluated = url
-        if not url_evaluated.startswith('http'):
-            url_evaluated = ''.join(['http://', url_evaluated])
-
-        url_parsed = urlparse(url_evaluated)
-        if 'resource_ssm_path' in url_evaluated:
-            self._get_url_params_from_query_key(url_params, url_parsed.query)
-        else:
-            self._get_url_params_from_url_path(url_params, url_parsed.path)
-
-        return url_params
-
-    def _get_url_params_from_query_key(self, url_params, query):
-        resource_path = dict(parse_qsl(query)).get('resource_ssm_path', '')
-        match = re.search(REGEX_OPAC_ALPHA_JOURNAL_ARTICLE_MEDIA_ASSETS_DETAILS, resource_path)
-        if match:
-            url_params['journal_acronym'] = match.group(1)
-            url_params['year_vol_issue'] = match.group(2)
-            url_params['file'] = match.group(3)
-            url_params['resource_ssm_path'] = resource_path
-
-            if query.endswith('.pdf'):
-                url_params['media_format'] = MEDIA_FORMAT_PDF
-
-    def _get_url_params_from_url_path(self, url_params, url_path):
-        for k, v in {
-            MEDIA_FORMAT_HTML: REGEX_OPAC_ALPHA_JOURNAL_ARTICLE_HTML_DETAILS,
-            MEDIA_FORMAT_PDF: REGEX_OPAC_ALPHA_JOURNAL_ARTICLE_PDF_DETAILS
-        }.items():
-            match = re.search(v, url_path)
-
-            if match:
-                url_params['media_format'] = k
-                url_params['journal_acronym'] = match.group(1)
-                url_params['year_vol_issue'] = match.group(2)
-                url_params['pages'] = match.group(3)
-                url_params['media_language'] = match.group(4)
-                return
-
-        path_match_assets = re.match(REGEX_OPAC_ALPHA_JOURNAL_ARTICLE_MEDIA_ASSETS_DETAILS, url_path)
-        if path_match_assets:
-            url_params['journal_acronym'] = path_match_assets.group(1)
-            url_params['year_vol_issue'] = path_match_assets.group(2)
-            url_params['file'] = path_match_assets.group(3)
-
-            if url_path.endswith('.pdf'):
-                url_params['media_format'] = MEDIA_FORMAT_PDF
-
-    def extract_media_format(self):
-        return self.url_params.get('media_format')
-
-    def extract_media_language(self, pid_v3):
-        media_language = self.url_params.get('media_language')
-        if not media_language:
-            media_language = self.articles_metadata['pid_v3_to_default_lang'].get(pid_v3, MEDIA_LANGUAGE_UNDEFINED)
-        return media_language
-
-    def extract_pid_v3(self):
-        return self._get_ssp_pid()
-    
-    def _get_ssp_pid(self):
-        artificial_pid = ':'.join([
-            self.url_params.get('journal_acronym', ''),
-            self.url_params.get('year_vol_issue', '')
-        ])
-
-        if 'pages' in self.url_params:
-            return ':'.join([artificial_pid, self.url_params.get('pages', '')])
-
-        if 'file' in self.url_params:
-            return ':'.join([artificial_pid, self.url_params.get('file', '')])
-    
-    def extract_issn(self):
-        return self.journals_metadata['acronym_to_scielo_issn'].get(self.url_params.get('journal_acronym'))
-
 
 class URLTranslatorOPACSite:
     def __init__(self, journals_metadata, articles_metadata):
